Compile to PostScript first
If we compile straight to PDF, it could be that all text will be
rasterized early on, which can be okay in many occasions. However, if
the goal is to send the document to be printed by an actual printing
press, it's best to use PostScript as an intermediate format, which
will prevent the text from being rasterized.

A result of this is that, at least for now, we must remove UoL Arms
image. It may very well be that it wasn't necessary anyway.

Signed-off-by: Felipe Balbi <felipe@balbi.sh>