Add support for subfiles
The subfiles packages will make it easier to handle a document that
can grow rather large. Instead of recompiling the entire document
every time, we can focus on a single chapter and recompile that single
chapter as a standalone document. Once all chapters are finalized,
then we can compile the entire document for final review.

Signed-off-by: Felipe Balbi <felipe@balbi.sh>